backend/loan-pool/requirements.txt
# Add requirements as necessary
# boto3 is present in lambda environment by default

pyjwt